requirements.txt
streamlit
pymongo
bcrypt
python-dotenv
groq
pandas
openpyxl
reportlab
matplotlib
streamlit-option-menu